+++ main.rb
@@ -1,7 +1,8 @@
 require './app'
-require './createPerson'
-require './createBook'
-require './createRental'
+require './create_person'
+require './create_book'
+require './create_rental'
+require './interractions'
 
 def options
   print "\nPlease choose an option by entering a number \n\n"
@@ -15,33 +16,10 @@ def options
   gets.chomp.to_i
 end
 
-def interraction(app)
-  user_input = options
-  case user_input
-  when 1
-    app.list_books
-    interraction(app)
-  when 2
-    app.list_people
-    interraction(app)
-  when 3
-    CreatePerson.new(app.people).create_person
-    interraction(app)
-  when 4
-    CreateBook.new(app.books).create_book
-    interraction(app)
-  when 5
-    CreateRental.new(app.rentals, app.books, app.people).create_rental
-    interraction(app)
-  when 6
-    app.list_rentals
-    interraction(app)
-  end
-end
-
 def main
   app = App.new
-  interraction(app)
+  interractions = Interraction.new
+  interractions.interraction(app)
 end
 
 main
